refactor(srv): split identity/resolver.rs into resolver/ submodules
Tier 3 item from docs/reports/REPORT_LARGE_FILE_MODULARIZATION_SCAN_2026_07_22.md.
Security-sensitive file — this module enforces SPEC_PROVIDER_ISOLATION's
INV-A (see docs/retro/retro-auth-isolation-invariant-silently-orphaned-
2026-07-14.md for the incident where it was silently orphaned once already).
inject_identity_env_with_broker (the credential-injection gate) and its full
21-test suite, including the INV-A regression canary test, were moved
together as one atomic unit into resolver/inject.rs, with the INV-A warning
doc comment copied verbatim and kept directly attached to the function (also
duplicated onto resolver/mod.rs's module doc as a belt-and-suspenders
measure). The peripheral, non-security-critical pieces (OAuth token-file
probing, error types, provider classification, secret resolution) split
cleanly into resolver/{oauth_probe,errors,provider,secret}.rs.

No logic changes — every function body, doc comment, and test assertion
copied verbatim. Test count matches exactly: 30 passed before and after
(independently re-verified after the agent's own report, plus a direct
manual re-read confirming the INV-A warning is intact and correctly
attached in its new location before this commit).

<!-- agentmux:agent_id=agentx -->
